refactor(MainPage): separated logics in custom hooks
